requirements.txt
Django==5.0.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-dateutil==2.8.2
django-filter==24.1 
dotenv==0.9.9
python-dotenv==1.1.0